parse in parallel
Currently parsing can be slow. To speed it up we chunk it, using rust's
available parallelism to find out how many chunks we can have.

in addition, the ip is not parsed anymore. When rendering it will always
use `0.0.0.0` to avoid that a remote adlist can have a malicious
mapping.